+++ components/support/error/src/lib.rs
@@ -174,4 +174,9 @@ macro_rules! define_error {
     };
 }
 
+#[cfg(feature = "glean-sym")]
+pub mod glean_metrics {
+    include!(concat!(env!("OUT_DIR"), "/glean_metrics.rs"));
+}
+
 uniffi::setup_scaffolding!("errorsupport");